// LocalAIAgent.WebUI/src/clients/UserApiClient/models/InvitationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InvitationDto = {
    id?: number;
    createdAt?: string;
    expiresAt?: string;
    redeemedAt?: string | null;
    revokedAt?: string | null;
    redeemedByUsername?: string | null;
};